requirements.txt
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
docx==0.2.4
exception==0.1.0
executing==2.0.1
Flask==3.0.0
fonttools==4.46.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonpickle==3.0.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
networkx==3.2.1
numpy==1.26.2
packaging==23.2
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.41
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
pyvis==0.3.2
PyYAML==6.0.1
six==1.16.0
stack-data==0.6.3
traitlets==5.14.0
typing_extensions==4.9.0
wcwidth==0.2.12
Werkzeug==3.0.1
